Make IDs FFI Safe
